+++ DTCD-StyleSystem/src/base-components/components.js
@@ -11,6 +11,11 @@ import BaseLabel from './BaseLabel/BaseLabel';
 import BaseLink from './BaseLink/BaseLink';
 import BaseIconButton from './BaseIconButton/BaseIconButton';
 import BaseChip from './BaseChip/BaseChip';
+import BaseTabs from './BaseTabs/BaseTabs';
+import BaseTooltip from './BaseTooltip/BaseTooltip';
+import BaseRadio from './BaseRadio/BaseRadio';
+import BaseRadioGroup from './BaseRadioGroup/BaseRadioGroup';
+import BaseRange from './BaseRange/BaseRange';
 
 export default [
   { name: 'base-button', baseClass: BaseButton },
@@ -24,6 +29,11 @@ export default [
   { name: 'base-link', baseClass: BaseLink },
   { name: 'base-icon-button', baseClass: BaseIconButton },
   { name: 'base-chip', baseClass: BaseChip },
+  { name: 'base-tabs', baseClass: BaseTabs },
+  { name: 'base-tooltip', baseClass: BaseTooltip },
+  { name: 'base-radio', baseClass: BaseRadio },
+  { name: 'base-radio-group', baseClass: BaseRadioGroup },
+  { name: 'base-range', baseClass: BaseRange },
   { name: 'datasource-select', baseClass: DataSourceSelect },
   { name: 'gauge-segment-builder', baseClass: GaugeSegmentBuilder },
 ];